requirements.txt
uvicorn==0.13.3
fastapi==0.65.1
starlette==0.14.2
pydantic==1.8.2
pydantic[email]==1.8.2
typing==3.7.4.3
mangum==0.11.0
SQLAlchemy==1.4.15
python-dotenv==0.15.0
boto3==1.17.78
pymysql==1.0.2
pyjwt==2.3.0
bcrypt==3.2.0
smtplib